% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse.smb_model}
\alias{analyse.smb_model}
\title{analyse.smb_model}
\usage{
\method{analyse}{smb_model}(x, data, parallel = getOption("mb.parallel",
  FALSE), quick = getOption("mb.quick", FALSE),
  quiet = getOption("mb.quiet", TRUE), glance = getOption("mb.glance",
  TRUE), beep = getOption("mb.beep", TRUE), ...)
}
\description{
analyse.smb_model
}
