% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R
\name{compare}
\alias{compare}
\title{Compare SMB analyses using WAIC}
\usage{
compare(analyses, ...)
}
\arguments{
\item{analyses}{An object inheriting from class smb_analyses.}

\item{...}{Not used.}
}
\description{
Compare SMB analyses using WAIC
}
