% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-block.R
\name{data_block}
\alias{data_block}
\title{Create Data Block from Data List}
\usage{
data_block(x)
}
\arguments{
\item{x}{A nlist}
}
\value{
A string
}
\description{
Automate writing the STAN data block from your data.
}
\examples{
mod_data <- nlist::as_nlist(list(
  X = c(1L, 2L, 3L, 4L),
  Y = c(1.2, 7.3, 8.9, 2.6),
  nObs = 4L
))
data_block(mod_data)
}
