% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R
\name{IC}
\alias{IC}
\title{IC Stan Analysis}
\usage{
IC(object, ...)
}
\arguments{
\item{object}{The smb_analysis object.}

\item{...}{Other arguments. Currently ignored.}
}
\value{
See \code{\link{waic}}
}
\description{
Widely applicable information criterion (WAIC)
}
\seealso{
\code{\link{waic}}
}
