% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.smb_analysis}
\alias{is.smb_analysis}
\title{Is a STAN Analysis}
\usage{
is.smb_analysis(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'smb_analysis'
}
