% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.smb_model}
\alias{is.smb_model}
\title{Is a STAN Model}
\usage{
is.smb_model(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'smb_model'
}
