% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.smb_analysis}
\alias{logLik.smb_analysis}
\title{Log-Likelihood}
\usage{
\method{logLik}{smb_analysis}(object, ...)
}
\arguments{
\item{object}{The smb_analysis object.}

\item{...}{unused.}
}
\description{
Log-Likelihood for a SMB analysis.
}
