% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo}
\alias{loo}
\title{LOO Stan Analysis}
\usage{
loo(analysis, ...)
}
\arguments{
\item{analysis}{The smb_analysis object.}

\item{...}{Optional arguments to pass to \code{\link{psislw}}. See \code{\link{loo}}.}
}
\value{
See \code{\link{loo}}
}
\description{
Leave-one-out cross-validation (LOO)
}
\details{
Efficient approximate leave-one-out cross-validation for Bayesian models. See
\link{loo-package} and Vehtari, Gelman, and Gabry (2016, 2017) for
background.
}
\seealso{
\code{\link{loo}}
}
