% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbr.R
\name{smb_dir}
\alias{smb_dir}
\title{List an SMB directory as a tibble}
\usage{
smb_dir(path)
}
\arguments{
\item{path}{SMB URL.}
}
\value{
A tibble with \code{name}, \code{type}, \code{smbc_type}, and \code{comment} columns.
}
\description{
List an SMB directory as a tibble
}
