% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbr.R
\name{smb_download}
\alias{smb_download}
\title{Copy an SMB file to the local filesystem}
\usage{
smb_download(remote, local, overwrite = FALSE)
}
\arguments{
\item{remote}{Source SMB URL.}

\item{local}{Destination local file path.}

\item{overwrite}{Whether to replace an existing local file.}
}
\description{
Copy an SMB file to the local filesystem
}
