% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbr.R
\name{smb_exists}
\alias{smb_exists}
\title{Test whether an SMB path exists}
\usage{
smb_exists(path)
}
\arguments{
\item{path}{SMB URL.}
}
\value{
A single logical value.
}
\description{
Test whether an SMB path exists
}
