% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbr.R
\name{smb_rename}
\alias{smb_rename}
\title{Rename an SMB file or directory}
\usage{
smb_rename(from, to)
}
\arguments{
\item{from}{Existing SMB URL.}

\item{to}{Destination SMB URL.}
}
\description{
Rename an SMB file or directory
}
