% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbr.R
\name{smb_upload}
\alias{smb_upload}
\title{Copy a local file to an SMB share}
\usage{
smb_upload(local, remote, overwrite = TRUE)
}
\arguments{
\item{local}{Local file path.}

\item{remote}{Destination SMB URL.}

\item{overwrite}{Whether to replace an existing remote file.}
}
\description{
Copy a local file to an SMB share
}
