% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbr.R
\name{smb_url}
\alias{smb_url}
\title{Construct and validate an SMB URL}
\usage{
smb_url(path, server = NULL, share = NULL)
}
\arguments{
\item{path}{A path beginning with \verb{smb://} or a server-relative path.}

\item{server}{Optional server used when \code{path} is relative.}

\item{share}{Optional share used when \code{path} is relative.}
}
\value{
A normalized SMB URL.
}
\description{
Construct and validate an SMB URL
}
