% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbr.R
\name{smb_write}
\alias{smb_write}
\title{Write raw data to an SMB file}
\usage{
smb_write(path, data, mode = "wb")
}
\arguments{
\item{path}{SMB URL.}

\item{data}{A raw vector, character vector, or object coercible to raw text.}

\item{mode}{File mode, normally \code{wb} or \code{ab}.}
}
\value{
Invisibly, the path.
}
\description{
Write raw data to an SMB file
}
