% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\title{WAIC Stan Analysis}
\usage{
waic(analysis, ...)
}
\arguments{
\item{analysis}{The smb_analysis object.}

\item{...}{Other arguments. Currently ignored.}
}
\value{
See \code{\link{waic}}
}
\description{
Widely applicable information criterion (WAIC)
}
\seealso{
\code{\link{waic}}
}
