% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic.smb_analysis}
\alias{waic.smb_analysis}
\title{WAIC Stan Analysis}
\usage{
\method{waic}{smb_analysis}(analysis, ...)
}
\arguments{
\item{analysis}{The smb_analysis object.}

\item{...}{Other arguments. Currently ignored.}
}
\value{
A named list (of class \code{'loo'}) with components:

\describe{
\item{\code{elpd_waic, se_elpd_waic}}{expected log pointwise predictive
density and standard error}
\item{\code{p_waic, se_p_waic}}{estimated effective number of parameters and
standard error}
\item{\code{waic, se_waic}}{\code{-2 * elpd_waic} (i.e., converted to the
deviance scale) and standard error}
\item{\code{pointwise}}{the pointwise contributions of each of the above
measures}
}
}
\description{
Widely applicable information criterion (WAIC)
}
\seealso{
\code{\link{waic}}
}
