% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-block.R
\name{data_block}
\alias{data_block}
\title{Create Data Block from Data List}
\usage{
data_block(data)
}
\arguments{
\item{data}{A list of parameters}
}
\value{
A string
}
\description{
Automate the data block straight from your data.
}
\examples{
mod_data <- list(
  X = c(1L, 2L, 3L, 4L),
  Y = c(1.2, 7.3, 8.9, 2.6),
  nObs = 4L
)
data_block(mod_data)
}
